requirements-dev.txt
-r requirements.txt

# Development only
django-extensions==3.2.3
django-debug-toolbar==4.1.0
black==23.11.0
flake8==6.1.0
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
